[FrameworkBundle][WebProfilerBundle] Move ProfilerPass to the WebProfiler bundle
