+++ src/Symfony/Bundle/FrameworkBundle/CHANGELOG.md
@@ -12,6 +12,7 @@ CHANGELOG
    is disabled.
  * Added `GlobalVariables::getToken()`
  * Deprecated `Symfony\Bundle\FrameworkBundle\DependencyInjection\Compiler\AddConsoleCommandPass`. Use `Symfony\Component\Console\DependencyInjection\AddConsoleCommandPass` instead.
+ * Deprecated `Symfony\Bundle\FrameworkBundle\DependencyInjection\Compiler\ProfilerPass`. Use `Symfony\Bundle\WebProfilerBundle\DependencyInjection\Compiler\ProfilerPass` instead.
 
 3.2.0
 -----
